DRYed up the Parser (10m)
