+++ lib/set_builder/trait.rb
@@ -1,11 +1,13 @@
 require "set_builder/constraint"
 require "set_builder/modifier"
+require "set_builder/parser"
 
 
 module SetBuilder
   class Trait
     attr_reader :expression, :tokens, :name, :modifiers, :direct_object_type, :enums
 
+    include Parser
 
 
     def initialize(expression, &block)
@@ -62,39 +64,6 @@ module SetBuilder
     def find(token)
       find_all(token).first
     end
-
-
-
-    def parse(expression)
-      tokenizer = Regexp.union(LEXER.values)
-      expression.split(tokenizer).each_with_object([]) do |lexeme, output|
-        next if lexeme.empty?
-        token = token_for(lexeme)
-        output.push [token, value_for(token, lexeme)]
-      end
-    end
-
-    def token_for(lexeme)
-      LEXER.each { |token, pattern| return token if pattern.match(lexeme) }
-      :string
-    end
-
-    def value_for(token, lexeme)
-      case token
-      when :name then lexeme[1...-1]
-      when :enum then lexeme[1...-1].split("|")
-      when :modifier then lexeme[1...-1]
-      when :direct_object_type then lexeme[1..-1]
-      else lexeme
-      end
-    end
-
-    LEXER = {
-      name: /("[^"]+")/,
-      direct_object_type: /(:[\w\-\.]+)/,
-      enum: /(\[[^\]]+\])/,
-      modifier: /(<\w+>)/
-    }.freeze
